requirements.txt
Django==4.2.10
Pillow==10.1.0
python-decouple==3.8
gunicorn==21.2.0
psycopg2-binary==2.9.7
whitenoise==6.6.0
